Release/285.0.0 (#5173)
## Explanation

This release will expose the first iteration of
token-search-discovery-controller v1.0.0.

## Reviewing Release PRs

- [x] Are there any packages with version bumps that should not be
published?
- For all packages being published:
- [x] Does the new version string sufficiently communicate the impact of
changes?
  - [x] Are there any changelog entries that do not need to be listed?
- [x] Are there any changelog entries that are placed in the wrong
category (including "Unreleased")?
- [x] Do all changelog entries explain the changes in a clear and
straightforward manner and detail exact changes to the package's
interface?
- [x] Do all change entries include a link to the pull requests where
that change was made?
- [x] Are there any changes since the package's previous release that
are not reflected in the changelog?
- [x] Have there been any changes merged to `main` while the release PR
has been open that have not been added to a changelog?
- [x] Are there any bumps to workspace dependencies made within the
release PR that have not been captured in the changelog?

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/token-search-discovery-controller`

- **Added**: More detail of the initial release in the changelog

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes