createServicePolicy: Convert callbacks to methods
Both RetryPolicy and CircuitBreakerPolicy from the Cockatiel library
allow for listening for events using `on*` methods. Our "service policy"
allows for listening to some of the same events as well as the
"degraded" event. For parity with Cockatiel — and for easier use in
general — this commit removes the callbacks that `createServicePolicy`
takes and replaces them with `on*` methods.