feat: SOL-46 adds the multichain transactions controller package (#5133)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

This PR adds a new package to the Core repo, with the new
MultichainTransactions controller, it comes from my original [extension
PR](https://github.com/MetaMask/metamask-extension/pull/29129), and it's
responsibility is to track transactions for non-EVM accounts, with
different block times, by using the Solana and Bitcoin snaps as the data
source. It shares the same approach/structure with the
MultichainBalancesController, given that responsibilities are very
similar.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

It relates to the [PR opened in
extension](https://github.com/MetaMask/metamask-extension/pull/29129)
that will be updated after this one is merged.

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/multichain-transactions-controller`

- **ADDED**: New `multichain-transactions-controller` package, with the
`MultichainTransactionsController` to track the transactions for non-EVM
accounts by using the Solana and Bitcoin snaps as the data source..

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes