Complete `createServicePolicy` (#5149)
This commit is a continuation of a previous commit which added a utility
function for use in service classes which ensures that they follow the
circuit breaker and retry patterns.

This commit completes the function by adding:

- `retryFilterPolicy`, which allows for retrying certain errors and not
handling others.
- `maxRetries`, which allows for giving up early before the max
consecutive failures is reached. Note that because this can combine with
`maxConsecutiveFailures`, it adds a bunch of other code paths and so
this requires adding a lot of new tests.