Add conveniences to createServicePolicy
Expose these exports from `cockatiel` so consumers can use them without
needing to depend on the library directly:

- `BrokenCircuitError`
- `CircuitState`
- `handleAll`
- `handleWhen`

Also, update `createServicePolicy` to expose the retry and circuit
breaker policies so that we can access their state in other modules that
use this utility function.

This is needed to be able to create an RPC service.