Fix NetworkController Redux selectors
When running tests, we are seeing the following warning:

> An input selector returned a different result when passed same
> arguments. This means your output selector will likely run more
> frequently than intended.

This is happening because the "input selector" in
`selectAvailableNetworkClientIds` returns a new object reference (returned
by `getNetworkConfigurations`) each time it is called. This violates the
Redux guidelines outlined here:
<https://redux.js.org/usage/deriving-data-selectors#optimizing-selectors-with-memoization>.

To fix this, we split off the "input selector" into its own selector
function which we then memoize.