Fix pre-push hook to fail if working tree dirty
Now that we've instructed the lint pipeline to update the ESLint warning
thresholds file, if you introduce changes to lint violations prior to
pushing, and you have the `pre-push` Git hook installed, you could
forget to push changes to this file.

This commit extracts the "working tree dirty" check that we perform in
the `lint-build-test` GitHub workflow to a script so that we can not
only call it in the workflow but also after running `yarn lint` in the
`pre-push` hook.

Note that you'll need to re-run `yarn simple-git-hooks` for this change
to take effect.