feat(accounts-controller): re-publish Snap keyring events (#5190)
## Explanation

The `SnapKeyring` now forward some keyring events coming from a Snap. To
avoid having controllers to depend the `SnapKeyring` for those new
events, we subscribe to them from the `AccountsController` and
re-publish them from it. This way, the `AccountsController` is the sole
dependency for those events (and is already a dependency of many
controllers anyway).

## References

- https://github.com/MetaMask/accounts/pull/154/files

## Changelog

### `@metamask/accounts-controller`

- **ADDED**: Re-publish Snap keyring events

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes