feat: add logo url to token search api service and update searchTokens path (#5195)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

The `/tokens-search/name` endpoint has changed to `/tokens-search` to
better reflect it's usecase in
[PR-75](https://github.com/consensys-vertical-apps/va-mmcx-portfolio-api/pull/75)
in the portfolio api. The endpoint searches by both name and address of
tokens from Moralis. Simply "/name" was misleading.

We also updated the param "name" to "query" to better reflect the nature
of the request.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/token-search-discover-controller`

- **Added**: The logoUrl is now exposed in TokenSearchResponseItem via
`searchTokens`
- **Breaking**:
- Update the TokenSearchApiService to use the updated URL for
`searchTokens`
    - The URL is now `/tokens-search` instead of `/tokens-search/name`
  - Changed the "name" parameter to "query" in the `searchTokens` method
- These updates align with the Portfolio API's `/tokens-search` endpoint
- The old endpoint `/tokens-search/name` will remain in a deprecated
state until parity across apps is achieved which for now is the
portfolio-api and core repos. The mobile
[PR-13111](https://github.com/MetaMask/metamask-mobile/pull/13111) is
not yet merged.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes