Fix NetworkController to handle clearing of messenger (#5116)
If the network controller is in the process of executing the
`lookupNetwork` step and the messenger is cleared of subscriptions, then
it may throw an error that the `networkDidChange` subscription is
missing. This happens in Mobile when it destroys the engine. There are
actually two places where `lookupNetwork` subscribes to
`networkDidChange` and then unsubscribes, but the aforementioned error
is only replicable with one of them. However, we handle both cases just
in case.