Add failover support to RpcService (#5202)
Add a new `failoverService` option to RpcService. This allows consumers
to forward the request to another RPC service in the event that an RPC
endpoint is proven to be unreliable (i.e. after attempting the request 5
times, the endpoint is unreachable, returns a a non-200 response, or
returns a response that is not JSON-decodable).