Ensure ESLint warning thresholds file stays updated (#5250)
We have a check in the main GitHub workflow where it will halt if, after
running `yarn lint`, it sees new changes show up in the working tree. At
the same time, in the new ESLint script, we have a check such that when
it is run on CI, it does not attempt to save updates to the ESLint
warning thresholds file if the number of warnings goes down, thereby
keeping the working tree clean and allowing CI to continue. This logic
was added intentionally because at the time some developers were
experiencing discrepancies in lint violations they saw locally vs. on
CI, and we didn't want the ESLint script to prevent these developers
from merging PRs.

However, there are two problems with this strategy:

1. It does not force developers to commit all changes to the warning
thresholds file, which means it may fall out of date. We end up having
to re-run `yarn lint` locally and apply any updates every so often for
this reason.
2. This strategy handles the case in which more violations show up
locally vs. on CI, but it doesn't handle when there are more violations
in CI vs. locally (which we've run into recently). So it's a
half-solution.

At this point, it seems like discrepancy issues happen less than before,
and we'd like to know every instance where it occurs rather than hide
some cases so we can fix them.

To that end, this commit removes the aforementioned check. This way,
when there are _any_ changes to the number of violations, the
developer is forced to commit updates to the file. This commit also
adds colors to the output to advise the user more clearly.