Add ability to create RPC service chain (#5226)
We want to update the `fetch` middleware in `eth-json-rpc-middleware`
and the Infura middleware in `eth-json-rpc-infura` to automatically fail
over to alternate RPC endpoints when the desired endpoint is down. To do
this, we need a way to string together a collection of RPC services,
where the first service represents the primary endpoint, the second
service is the failover for the first, the third service is a failover
for the second, etc. This composite object should conform to the same
interface as an RPC service so consumers do not have to care whether
they are working with a single endpoint or a chain of them.