Ensure core files are codeowned by Wallet Framework
I noticed that when new packages are added we may not always know about
it. It seems that there are many files that are not codeowned by any
team and so anyone is free to update them. This is especially true of
the TypeScript config files but also the README.

I've added the `github-codeowner` package as a dev dependency to ensure
that we can stay on top of this better. You can run this command to find
all unowned files:

```
yarn github-codeowner -u
```

And if you're unsure who owns a file, you can also run:

```
yarn github-codeowner who <file>
```

With that context, I've gone through and granted as many files to Wallet
Framework as made sense.