Fix stray console.error in NetworkController tests
When running one of the tests for the `destroy` method, an error from
the block tracker can be observed. This error does not fail the test,
but it creates noise in the output. It turns out that this happens
because we are failing to mock a request in the test. This commit adds
the requisite test so that the `console.error` goes away.