RpcService: Regard node-fetch errors as retriable
We use `node-fetch` internally for testing. This library produces its
own error messages whenever a request cannot be initiated, so we need to
account for it within the RpcService retry logic. However, we need to
make sure to exclude errors from Nock so we don't get weird results in
tests.