Revert wildcard export in AccountsController
Using a wildcard export in the root `index.ts` of a package is not an
ideal practice. This file is very useful because it gives us a way to
understand everything that a package publicly exports at a glance.

In using a wildcard export, however, we lose this ability. Because all
symbols marked with `export` will automatically become exports of the
package, we can no see when a new package is introduced in a PR — so it
is possible to introduce a new export without even knowing about it —
and we lose the ability to export a symbol purely for internal purposes,
which is useful for tests.

As it relates to the change being reverted, because of the introduction
of this wildcard export, new exports have been introduced which will now
need to be reverted in a new major version, namely `AllowedActions` and
`AllowedEvents`. This is non-ideal; we should not have to introduce
major versions if we do not need to, because they are disruptive.

Earlier it was noted that needing to add an export manually to this list
is a small pain point. I would argue that not knowing about new exports
creates _more_ of a pain point in the long run. That is, this is a
feature and not a bug :)