chore: Change ownership of SelectedNetworkController
The package `@metamask/selected-network-controller` is now owned solely
by the Wallet API Platform team, rather than being shared between
multiple teams. This team has the most context for what this code does,
and it will in the future be increasingly less relevant to the rest of
the wallet apart from supporting the EIP-1193 provider.