RpcService: Regard node-fetch errors as retriable (#5298)
Currently we have tests for middleware to verify that requests get
retried when certain errors are thrown. In a future commit we will
replace the retry logic in these middleware with RpcService, and when
this happens the aforementioned tests will break. This is because we are
using Nock to force the request to throw errors, and in tests we use
`node-fetch` to polyfill the `fetch` function, and `node-fetch` errors
are not regarded as retriable by RpcService.

This commit adjusts RpcService to treat `node-fetch` errors as retriable
(but making sure to exclude errors from Nock itself, since they will
also manifest as `node-fetch` errors).