feat: Add `MultichainNetworkController` to handle both EVM and non-EVM network and account switching (#5215)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

This PR updates both the MultichainNetworkController and
AccountsController to handle network switching as well as account
switching. The logic handles the following logic:
- Switching accounts on AccountsController will notify
MultichainNetworkController to update if the account belongs to evm vs
non-evm network (MultichainNetworkController subscribes to
AccountsController event)
- Switching between networks on MultichainNetworkController will notify
AccountsController to update accounts based on which network the account
belongs to (AccountsController subscribes to MultichainNetworkController
event)



## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

Fixes https://github.com/MetaMask/accounts-planning/issues/804

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/accounts-controller`

- **BREAKING**:
- Added `MultichainNetworkController:networkDidChange` to allowed
events. This is used to subscribe to the `setActiveNetwork` event from
the `MultichainNetworkController` and is responsible for updating
selected account based on network changes (both EVM and non-EVM).

### `@metamask/multichain-network-controller`

- **ADDED**:  
- Allowed actions - `NetworkControllerGetStateAction` |
`NetworkControllerSetActiveNetworkAction`. The
`MultichainNetworkController` acts as a proxy for the
`NetworkController` and will update it based on EVM network changes.
- Allowed events - `AccountsControllerSelectedAccountChangeEvent` to
allowed events. This is used to subscribe to the `selectedAccountChange`
event from the `AccountsController` and is responsible for updating
active network based on account changes (both EVM and non-EVM).

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: gantunesr <17601467+gantunesr@users.noreply.github.com>
Co-authored-by: tommasini <tommasini15@gmail.com>
Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>
Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>