Create RPC middleware using RPC services
When creating the middleware stack for a network, NetworkController
makes use of functions from two packages for sending requests:

- `createInfuraMiddleware` from `@metamask/eth-json-rpc-infura` for Infura networks
- `createFetchMiddleware` from `@metamask/eth-json-rpc-middleware` for custom RPC endpoints

Currently, each of these middleware implements similar heuristics for
determining when the network is down and retrying requests
appropriately. We want to improve upon this logic so that we incorporate
the circuit breaker pattern, exponential backoff, and jitter for
retries, and we have implemented an RpcService class which extracts all
of this logic into one place. We can vastly simplify the Infura and
fetch middleware creation functions so that they only need to take an
RPC service and the RPC service will handle the rest.

To accommodate these changes, this commit:

- Upgrades the `@metamask/eth-json-rpc-infura` and
  `@metamask/eth-json-rpc-middleware` packages to allow an RPC service
  to be passed.
- Adds two new required options to NetworkController, `fetch` and
  `btoa`. These are also requirements of RpcService and allows consumers
  to pass in whatever versions of these functions is specific to their
  platform.
- Updates `createAutoManagedNetworkClient` to take `fetch` and `btoa`
  and pass them along to `createNetworkClient`.
- Updates `createNetworkClient` to create an RPC service and pass it
  along to the two middleware creation functions. Consequently, these
  functions now handle requests in exactly the same way.