fix: fix account controller legacy polling (#5321)
## Explanation
Changes were made to accountTrackerController inside the
@metamask/assets-controllers package.

The account tracker polling was executed on the constructor with the
initialisation. This caused additional requests to be sent by the app,
leading to performance issues.

Additionally, the accountTracker was still using legacy polling via the
poll function. This function has now been removed, and a test has been
added to prevent the creation of multiple polling instances.

These changes improve efficiency by reducing unnecessary network
requests and ensuring the polling mechanism is properly controlled.

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **CHANGED**: Removed legacy poll function to prevent redundant
polling.
- **FIXED**: ensure that the polling is not triggered on the constructor
with the initialisation


## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

Co-authored-by: sahar-fehri <sahar.fehri@consensys.net>