Fix findNetworkClientIdByChainId to consider default endpoints
Currently, the `findNetworkClientIdByChainId` method in
`NetworkController`, which is used in various places, mostly commonly
TransactionController and the wallet middleware, returns the network
client ID corresponding to the first listed RPC endpoint for the given
chain. This may not match users' expectations if they have chosen
another RPC endpoint as the default. This commit fixes the
implementation of `findNetworkClientIdByChainId` to match this
expectation.