Support keyring metadata in KeyringController (#5112)
## Explanation

Currently, we assume that the only HD Keyring is the Primary Keyring.
Most methods related to adding accounts or displaying the SRP assume
that it refers to the first keyring, which is the Primary Keyring. To
enable having more than one SRP, and consequently more than one HD
Keyring, this PR adds an array called `keyringsMetadata`, which contains
an id that can be used to identify all keyrings and a proper place to
keep other data we might want to add to each keyring such as `name`.

`keyringsMetadata` is a separate in order to minimize changes that we
need to make to keyrings and the way we process them. We also can keep
this data outside of the vault, as it's not storing any sensitive data.

## References

Related to [ADR
0002-keyring-id-and-name.md](https://github.com/MetaMask/decisions/blob/keyring-id-and-name/decisions/accounts/0002-keyring-id-and-name.md)
Blocks [Extension
PR](https://github.com/MetaMask/metamask-extension/pull/29794)
## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/keyring-controller`

- **ADDED**: `{ id: string }` selector to `withKeyring` function
- **ADDED**: optional `keyringId` param to `addNewAccount` function
- **ADDED**: optional `keyringId` param to `exportSeedPhrase` function
- **ADDED**: optional `keyringId` param to `getAccounts` function
- **ADDED**: optional `keyringId` param to `verifySeedPhrase` function
- **ADDED**: private `#getKeyringById` helper function
- **ADDED**: optional `keyringId` param to `#verifySeedPhrase` function
- **ADDED**: optional `metadata` param to `#newKeyring` function
- **ADDED**: optional `metadata` param to `#newKeyring` function

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: Michele Esposito <34438276+mikesposito@users.noreply.github.com>
Co-authored-by: Monte Lai <monte.lai@consensys.net>
Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>
Co-authored-by: Shane Terence Odlum <shane.odlum@consensys.net>
Co-authored-by: Mark Stacey <markjstacey@gmail.com>